fullstackAI/jobs/requirements.txt
requests==2.31.0
pandas==2.2.0
numpy==1.26.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
yfinance==0.2.36
schedule==1.2.1
redis==5.0.1
sqlalchemy==2.0.25
dagster
dagster-webserver
